+++ packages/sources/react/src/components/forms/VtmnSelect/VtmnSelect.tsx
@@ -1,12 +1,10 @@
 import React from 'react';
 import clsx from 'clsx';
 import '@vtmn/css-select/dist/index-with-vars.css';
-import { objectValuesToString } from '../../../utils/object';
 
 export interface VtmnSelectProps extends React.ComponentPropsWithRef<'select'> {
   error?: boolean;
   errorText?: string;
-  helperText?: string;
   id: string;
   labelText?: string;
   options: React.ComponentPropsWithoutRef<'option'>[];
@@ -35,7 +33,7 @@ export const VtmnSelect = ({
           'vtmn-select--error': error,
         })}
         aria-describedby={hasErrorText ? errorTextId : undefined}
-        {...objectValuesToString(props)}
+        {...props}
       >
         {options.map((option) => option)}
       </select>