+++ packages/sources/react/src/components/indicators/VtmnTag/index.ts
@@ -1,0 +1,2 @@
+export { default as VtmnTag } from './VtmnTag';
+export * from './VtmnTag';