+++ packages/sources/svelte/src/components/actions/VtmnButton/test/VtmnButtonWithSlots.svelte
@@ -1,0 +1,7 @@
+<script>
+  import VtmnButton from '../VtmnButton.svelte';
+</script>
+
+<VtmnButton on:click {...$$restProps}>
+  <span data-ut="slot"> Unit-button </span>
+</VtmnButton>