+++ packages/sources/svelte/src/components/overlays/VtmnToast/VtmnToastItem.svelte
@@ -25,7 +25,7 @@
    */
   export let timeout;
 
-  let className = '';
+  let className = undefined;
   /**
    * @type {string} Custom classes to apply to the component.
    */