+++ packages/sources/svelte/src/components/structure/VtmnAccordion/test/VtmnAccordionWithSlots.svelte
@@ -1,0 +1,5 @@
+<script>
+  import VtmnAccordion from '../VtmnAccordion.svelte';
+</script>
+
+<VtmnAccordion {...$$restProps}>Unit test content</VtmnAccordion>