+++ packages/sources/svelte/src/components/structure/VtmnListItem/VtmnListItem.svelte
@@ -1,4 +1,7 @@
 <script>
+  import { VTMN_LISTITEM_SIZE } from './enums';
+  import { cn } from '../../../utils/classnames';
+
   /** @restProps */
 
   const SLOTS = $$props.$$slots;
@@ -9,7 +12,7 @@
    * @type {'small' | 'medium'} [medium]
    * @defaultValue 'medium'
    */
-  export let size = 'medium';
+  export let size = VTMN_LISTITEM_SIZE.MEDIUM;
 
   /**
    * Has divider.
@@ -20,19 +23,24 @@
   export let divider = true;
 
   /**
-   * Index of item in list.
-   *
-   * @type {number} item
-   */
-  export let index;
-
-  /**
    * Set disabled state of list item.
    *
    * @type {boolean} [disabled]
    * @defaultValue false
    */
   export let disabled = false;
+
+  let className = undefined;
+  /**
+   * @type {string} Custom classes to apply to the component.
+   */
+  export { className as class };
+
+  $: componentClass = cn(
+    `vtmn-list_item-size--${size}`,
+    !divider && 'vtmn-list_item--without-divider',
+    className,
+  );
 
   /**
    * Delete slot div from de DOM if slot doesn't exist.
@@ -47,10 +55,9 @@
 
 <li
   on:click
-  class={`vtmn-list_item-size--${size}`}
-  class:vtmn-list_item--without-divider={!divider}
+  class={componentClass}
   role="option"
-  tabindex={index}
+  tabindex="0"
   aria-disabled={disabled}
   {...$$restProps}
 >