+++ packages/sources/vue/src/components/index.ts
@@ -5,10 +5,12 @@ import VtmnSelect from './forms/VtmnSelect';
 import VtmnBadge from './indicators/VtmnBadge';
 import VtmnLoader from './indicators/VtmnLoader';
 import VtmnPrice from './indicators/VtmnPrice';
+import VtmnTag from './indicators/VtmnTag';
 import VtmnCheckbox from './selection-controls/VtmnCheckbox';
 import VtmnChip from './selection-controls/VtmnChip';
 import VtmnRadioButton from './selection-controls/VtmnRadioButton';
 import VtmnToggle from './selection-controls/VtmnToggle';
+import VtmnDivider from './structure/VtmnDivider';
 
 export {
   VtmnButton,
@@ -21,5 +23,7 @@ export {
   VtmnCheckbox,
   VtmnChip,
   VtmnRadioButton,
+  VtmnTag,
   VtmnToggle,
+  VtmnDivider,
 };