+++ packages/sources/vue/src/components/indicators/VtmnTag/index.ts
@@ -1,0 +1,8 @@
+import { App as Application } from 'vue';
+import VtmnTag from './VtmnTag.vue';
+
+VtmnTag.install = (Vue: Application) => {
+  Vue.component(VtmnTag.name, VtmnTag);
+};
+
+export default VtmnTag;