+++ packages/sources/vue/src/components/structure/VtmnDivider/index.ts
@@ -1,0 +1,8 @@
+import { App as Application } from 'vue';
+import VtmnDivider from './VtmnDivider.vue';
+
+VtmnDivider.install = (Vue: Application) => {
+  Vue.component(VtmnDivider.name, VtmnDivider);
+};
+
+export default VtmnDivider;