+++ packages/sources/vue/src/components/structure/VtmnDivider/types.ts
@@ -1,0 +1,3 @@
+export type VtmnDividerOrientation = 'horizontal' | 'vertical';
+
+export type VtmnDividerTextPosition = 'start' | 'center' | 'end';