+++ packages/sources/css/presets/tailwind/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.33.86](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.85...@vtmn/css-tailwind-preset@0.33.86) (2022-03-25)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
 ## [0.33.85](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.84...@vtmn/css-tailwind-preset@0.33.85) (2022-03-24)
 
 **Note:** Version bump only for package @vtmn/css-tailwind-preset